Add a code-staging directory
This commit adds a file path specification for a code staging directory.
The intent is for use with the file-sync service. r10k or other
automatic deployment utilities can prepare a new code version for
deployment here, prior to triggering an atomic commit action to deploy
the new code to the live codedir.